Separated the encryption and file access functions into their own files to allow for more modular code.
